Migrate tests to pytest
