--- almanac/tests/test_command_engine.py
@@ -1,8 +1,0 @@
-"""Tests for the ``CommandEngine`` class."""
-
-from unittest import TestCase
-
-
-class CommandEngineTestCase(TestCase):
-    # TODO
-    pass
