--- almanac/tests/test_token_parse.py
@@ -1,8 +1,0 @@
-"""Tests for token-parsing functionality."""
-
-from unittest import TestCase
-
-
-class TokenParsingTestCase(TestCase):
-    # TODO
-    pass
