+++ deps/dev-requirements.txt
@@ -9,6 +9,8 @@ pycodestyle==2.6.0
 pyflakes==2.2.0
 Pygments==2.6.1
 pyparsing==2.4.7
+pytest==6.0.1
+pytest-asyncio==0.14.0
 readme_renderer[md]==26.0
 Sphinx==3.1.2
 sphinx-autodoc-typehints==1.11.0
