+++ pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+addopts = --doctest-modules --doctest-glob "*.rst"
+testpaths = tests almanac docs
